fix: tweak CPV parsing rules to match PMS.
Bug #421 captures this; pkgcore was allowing version components
as the package name if it was suffixed by a revision; the fix
for that incorrectly limited a package name that has a trailing
revision, but *no version syntax* in the name.

This commit just refactors the revision check to also force a check
for a leading version if revision-like is found.

Resolves: https://github.com/pkgcore/pkgcore/issues/421
Signed-off-by: Brian Harring <ferringb@gmail.com>
Closes: https://github.com/pkgcore/pkgcore/pull/422
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>