fix filtered.tree.{categories,packages,versions} .
The implementation wasn't properly aliasing _get_* to
the underlying raw_repo, and it should've been filtering
packages and versions according to the filter.  This adds that.

This isn't a huge issue, but it breaks API guarantees
for code tring to efficienctly access this sort of info
for a repository stack.

Signed-off-by: Brian Harring <ferringb@gmail.com>