fix: tox.ini since it's not been used in ages.
This broken whenever the repo went to pyproject.toml .

On the one hand, I like enumerating what devs have to
invoke locally rather than relying on GH.

On the other hand, clearly no one is using this.

This fixes it for current usage.  If it's to continue,
then GH and tox envlist's should be fused.

Signed-off-by: Brian Harring <ferringb@gmail.com>