docs: make the manpages contain the extended help text
This uses the snake oil from snakeoil.cli.arghparse :) which passes
extra docs in via custom argparse attributes. In order to produce docs,
you need to set a snakeoil global variable, which then causes the extra
docs to replace the help?

This also revealed a bug. You cannot simply use the percent symbol in
argparse help strings, as documented in the standard library:

> As the help string supports %-formatting, if you want a literal % to
> appear in the help string, you must escape it as %%.

Indeed, it crashes if you include one such as ``%`` since the backtick
isn't a valid format specifier. Fix that by percent-escaping the
percent.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>