chore: import snakeoil.tar so pkgcore usage can be removed at our pace
snakeoil.tar existed to fix upstream python issues in memory usage; those
were fixed somewhere in the early py3k days, but snakeoil also eventually
added support for the snakeoil.data_sources protocol, which also is going
to be deprecated and removed.

This tar module is moved here because pkgcore binpkgs is the only thing
that uses it, so migration away from that will be handled within pkgcore
itself.

Signed-off-by: Brian Harring <ferringb@gmail.com>