force bash-5.3 for doc run.
It's arguable if we should require this- it's exploding due to
a hardwired check in the eapi instance code.  This makes sense
for any  possible EBD helpers that are pure EAPI9 (thus bash5.3),
but they don't exist.

Beyond that, the inability to import pkgcore when an EAPI deps
are missing explicitly precludes the ability to ever do merges to
resolve this.  The EAPI should instead be disabled, but that has
a metric shitton of downstream API consequences.

Signed-off-by: Brian Harring <ferringb@gmail.com>